pyproject.toml
[project]
name = "passview-chatbot-poc"
version = "0.1.0"
description = "Default template for PDM package"
authors = [{ name = "jjangga0214", email = "bnbcmindnpass@gmail.com" }]
dependencies = [
    "streamlit-chat>=0.1.1",
    "streamlit>=1.37.0",
    "langchain>=0.2.11",
    "faiss-cpu>=1.8.0.post1",
    "python-dotenv>=1.0.1",
    "langchain-community>=0.2.10",
    "pypdf>=4.3.1",
    "openai>=1.37.1",
    "tiktoken>=0.7.0",
    "unstructured>=0.15.0",
    "sentence-transformers>=3.0.1",
    "chromadb>=0.5.5",
]
requires-python = "==3.12.*"
readme = "README.md"
license = { text = "Copyright" }

[tool.pdm]
distribution = false
